+++ CHANGES.md
@@ -70,6 +70,9 @@
     of processing messages in `broadcastMessage`. Previously,
     `runOnWorkspaces` processed the hidden workspaces first.
 
+  * Added `withUnfocused` function to `XMonad.Operations`, allowing for
+    `X` operations to be applied to unfocused windows.
+
 ## 0.15 (September 30, 2018)
 
   * Reimplement `sendMessage` to deal properly with windowset changes made